services/image_caption_docker/python_requirements.txt
transformers==4.36.2
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
pandas
google-cloud-storage
pyarrow